requirements.txt
pip==21.3.1
setuptools==60.2.0
wheel==0.37.1

uvicorn~=0.22.0
torch~=2.0.1
fastapi~=0.98.0
Pillow~=9.5.0
torchvision~=0.15.2